+++ resources/views/edit-inventory.blade.php
@@ -208,9 +208,9 @@
                                     <div id="genres-field-{{$inventoryItem->id}}">
                                         <ul class="mb-0" id="genres-ul-{{$inventoryItem->id}}">
                                             @foreach($inventoryItem->genres as $genre)
-                                                <li>
-                                                    <span id="genre-{{$genre->id}}" class="genre">{{$genre->name}}</span>
-                                                    <button class="btn btn-sm btn-danger remove-genre-button">
+                                                <li class="genre-li" id="genre-{{$genre->id}}-for-{{$inventoryItem->id}}-li">
+                                                    <span>{{$genre->name}}</span>
+                                                    <button class="btn btn-sm btn-danger remove-genre-button" id="button-remove-genre-{{$genre->id}}-for-{{$inventoryItem->id}}">
                                                         <i class="fas fa-times"></i>
                                                     </button>
                                                 </li>